Create Japanese.json
So, this translation, is the latest version of the translation. If there is more update, then I would try to keep up with the update, along with the English version (English.json). Some (barely) of the translation, I used the version Rikku2002 made as a guideline, to make a more clear to understand Japanese version of the translation, so I have their name included on the author of the translation.